+++ src/main/java/org/cojen/tupl/_LocalDatabase.java
@@ -3194,7 +3194,7 @@ final class _LocalDatabase extends AbstractDatabase {
 
         /*P*/ // [|
         if (mFullyMapped) {
-            node.mPage = mPageDb.directPagePointer(node.mId);
+            node.mPage = mPageDb.dirtyPage(node.mId);
         }
         /*P*/ // ]
 
@@ -3376,7 +3376,7 @@ final class _LocalDatabase extends AbstractDatabase {
         /*P*/ // [|
         if (mFullyMapped) {
             if (node.mPage == p_nonTreePage()) {
-                node.mPage = mPageDb.directPagePointer(newId);
+                node.mPage = mPageDb.dirtyPage(newId);
                 node.asEmptyRoot();
             } else if (node.mPage != p_closedTreePage()) {
                 node.mPage = mPageDb.copyPage(node.mId, newId); // copy on write
@@ -3400,7 +3400,12 @@ final class _LocalDatabase extends AbstractDatabase {
      * Caller must hold commit lock and exclusive latch on node. This method
      * should only be called for nodes whose existing data is not needed.
      */
-    void redirty(_Node node) {
+    void redirty(_Node node) throws IOException {
+        /*P*/ // [|
+        if (mFullyMapped) {
+            mPageDb.dirtyPage(node.mId);
+        }
+        /*P*/ // ]
         mDirtyList.add(node, mCommitState);
     }