+++ src/test/java/org/cojen/tupl/LargeValueDirectMappedTest.java
@@ -31,6 +31,6 @@ public class LargeValueDirectMappedTest extends LargeValueTest {
     @Before
     @Override
     public void createTempDb() throws Exception {
-        mDb = TestUtils.newTempDatabase(80_000_000L, TestUtils.OpenMode.DIRECT_MAPPED);
+        mDb = TestUtils.newTempDatabase(200_000_000L, TestUtils.OpenMode.DIRECT_MAPPED);
     }
 }